Oracle wrapper packs any component count; Image channels widen to u16
The openjpeg wrapper cast numcomps to u8 (a panic at the 257-component
conformance file p0_13) and hit `unreachable` on 2 components (p1_07).
A valid file must never crash the oracle. The wrapper now bounds the
count by Csiz (16384), packs every count, and names 1/3/4 with their
layouts and everything else as the new PixelLayout.multichannel.

Image.channels and ImageMetadata.channels are u16; the C ABI follows
(uint16_t channels, JP2Z_LAYOUT_MULTICHANNEL) and the CLI writes PAM
(P7, DEPTH = channels) for counts other than 1 and 3 instead of
refusing. jp2z's own decode maps 4 channels to cmyk and any other
count to multichannel; its 16-component ceiling is the next item.

RED test: p0_13 and p1_07 through internal.openjpegDecode, asserting
channel count, layout, precision and pixel length; it aborted with
"integer does not fit in destination type" before the fix.

+++ src/ffi/c_api.zig
@@ -47,7 +47,7 @@ const CImage = extern struct {
     pixels_len: usize,
     width: u32,
     height: u32,
-    channels: u8,
+    channels: u16,
     bits_per_sample: u8,
     source_color_space: c_int,
     layout: c_int,
